# packages/shared-types/openapi.yaml
openapi: 3.1.0
info:
  title: VIGIL API
  version: 1.0.0
  summary: Distress-monitoring API for survivors of atrocities (SIH26094).
  description: |
    Source of truth for the frontend ↔ gateway ↔ service contract. TypeScript
    types are generated from this file (`pnpm contract:generate`); do not
    hand-write them.

    **Auth.** Every operation requires a signed RS256 JWT whose `scope` claim is
    `victim`, `counselor` or `admin`. `x-vigil-scopes` on each operation is read
    by the gateway at boot and enforced there — `/v1/admin/*` is rejected at the
    gateway for anything but `admin`. `victim` tokens may only touch their own
    `victimId` (the token's `sub`); services enforce ownership.

    **Dignity rules baked into the contract.** `Alert.raw_score` is always
    `null`. Free text is never returned by list endpoints unless the victim's
    consent scope `share-free-text` was granted, or it arrived on the crisis
    fast-path.
servers:
  - url: https://localhost:8080
    description: Local gateway (docker compose)
security:
  - bearerAuth: []
...
paths:
...
  /v1/victims/{id}/trend:
    get:
      tags: [trend]
      operationId: getVictimTrend
      x-vigil-service: distress-trend-service
      x-vigil-scopes: [counselor]
      parameters:
        - $ref: '#/components/parameters/VictimIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrendResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/scoring-config:
    get:
      tags: [trend]
      operationId: getScoringConfig
      x-vigil-service: distress-trend-service
      x-vigil-scopes: [victim, counselor, admin]
      summary: Current versioned scoring config (weights + thresholds).
      description: Devices cache this and stamp every check-in with `configVersion`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScoringConfig' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /v1/reminders:
    get:
      tags: [reminders]
      operationId: listReminders
      x-vigil-service: reminder-service
      x-vigil-scopes: [victim, counselor]
      parameters:
        - name: victimId
          in: query
          required: true
          schema: { $ref: '#/components/schemas/VictimId' }
      responses:
        '200':
          description: OK, ordered by dueAt ascending
          content:
            application/json:
              schema:
                type: object
                required: [reminders]
                additionalProperties: false
                properties:
                  reminders:
                    type: array
                    items: { $ref: '#/components/schemas/Reminder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [reminders]
      operationId: upsertReminder
      x-vigil-service: reminder-service
      x-vigil-scopes: [counselor]
      summary: Create or update a reminder (server-authoritative updatedAt, last-write-wins).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReminderInput' }
      responses:
        '201':
          description: Created or updated
          content:
            application/json:
              schema:
                type: object
                required: [reminder]
                additionalProperties: false
                properties:
                  reminder: { $ref: '#/components/schemas/Reminder' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /v1/admin/rollups:
    get:
      tags: [admin]
      operationId: getAdminRollups
      x-vigil-service: distress-trend-service
      x-vigil-scopes: [admin]
      summary: Anonymized regional aggregate (k-anonymity suppressed below cohort 5).
      parameters:
        - name: region
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [anonymizedAggregate]
                additionalProperties: false
                properties:
                  anonymizedAggregate: { $ref: '#/components/schemas/AnonymizedAggregate' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    VictimIdPath:
      name: id
      in: path
      required: true
      schema: { $ref: '#/components/schemas/VictimId' }
  responses:
    BadRequest:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Token scope not permitted for this operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, message]
      additionalProperties: false
      properties:
        error: { type: string }
        message: { type: string }
    VictimId:
      type: string
      description: Pseudonymous identifier. Never a name, phone number or case number.
      pattern: '^v_[A-Za-z0-9_-]{6,64}$'
    Timestamp:
      type: integer
      description: Unix epoch milliseconds
      minimum: 0
...
    Baseline:
      type: object
      required: [mean, stddev, windowSize, sufficient]
      additionalProperties: false
      properties:
        mean: { type: number }
        stddev: { type: number, minimum: 0 }
        windowSize: { type: integer, minimum: 0 }
        sufficient:
          type: boolean
          description: False until the minimum number of check-ins exists; no trend alert can fire before then.
    TrendPoint:
      type: object
      required: [at, compositeScore]
      additionalProperties: false
      properties:
        at: { $ref: '#/components/schemas/Timestamp' }
        compositeScore: { type: number, minimum: -1, maximum: 1 }
    TrendResponse:
      type: object
      required: [points, baseline]
      additionalProperties: false
      properties:
        points:
          type: array
          items: { $ref: '#/components/schemas/TrendPoint' }
        baseline:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/Baseline'
    ScoringConfig:
      type: object
      required: [version, weights, zThreshold, urgentZThreshold, slopeThreshold, baselineWindow, minBaseline, slopeWindow, sigmaFloor, involvementThreshold]
      additionalProperties: false
      properties:
        version: { type: integer, minimum: 1 }
        weights:
          type: object
          required: [selfReport, sentiment, engagement, voice]
          additionalProperties: false
          properties:
            selfReport: { type: number, minimum: 0, maximum: 1 }
            sentiment: { type: number, minimum: 0, maximum: 1 }
            engagement: { type: number, minimum: 0, maximum: 1 }
            voice: { type: number, minimum: 0, maximum: 1 }
        zThreshold: { type: number, maximum: 0, description: 'T — distressed direction is negative' }
        urgentZThreshold: { type: number, maximum: 0 }
        slopeThreshold: { type: number, exclusiveMinimum: 0, description: Magnitude; a slope below -slopeThreshold is distressed. }
        baselineWindow: { type: integer, minimum: 2 }
        minBaseline: { type: integer, minimum: 2 }
        slopeWindow: { type: integer, minimum: 2 }
        sigmaFloor: { type: number, exclusiveMinimum: 0 }
        involvementThreshold: { type: number, exclusiveMinimum: 0 }
...
    ReminderType:
      type: string
      enum: [hearing, compensation, checkin]
    ReminderInput:
      type: object
      required: [id, victimId, dueAt, type, note]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        victimId: { $ref: '#/components/schemas/VictimId' }
        dueAt: { $ref: '#/components/schemas/Timestamp' }
        type: { $ref: '#/components/schemas/ReminderType' }
        note: { type: string, maxLength: 500 }
    Reminder:
      type: object
      required: [id, victimId, dueAt, type, note, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        victimId: { $ref: '#/components/schemas/VictimId' }
        dueAt: { $ref: '#/components/schemas/Timestamp' }
        type: { $ref: '#/components/schemas/ReminderType' }
        note: { type: string, maxLength: 500 }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    AnonymizedAggregate:
      type: object
      required: [region, periodDays, suppressed, cohortSize, checkInCount, alerts, weeklyCheckIns]
      additionalProperties: false
      properties:
        region: { type: string }
        periodDays: { type: integer, minimum: 1 }
        suppressed:
          type: boolean
          description: True when the cohort is below the k-anonymity threshold; all counts are then null.
        cohortSize: { type: [integer, 'null'], minimum: 0 }
        checkInCount: { type: [integer, 'null'], minimum: 0 }
        alerts:
          oneOf:
            - type: 'null'
            - type: object
              required: [urgent, elevated, acknowledgedWithin24hPct]
              additionalProperties: false
              properties:
                urgent: { type: integer, minimum: 0 }
                elevated: { type: integer, minimum: 0 }
                acknowledgedWithin24hPct: { type: [number, 'null'], minimum: 0, maximum: 100 }
        weeklyCheckIns:
          type: array
          items:
            type: object
            required: [weekStart, count]
            additionalProperties: false
            properties:
              weekStart: { $ref: '#/components/schemas/Timestamp' }
              count: { type: integer, minimum: 0 }